requirements.txt
# Core ML and Data Science
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Chemistry and Molecular Fingerprints
rdkit-pypi>=2022.9.1
mordred>=1.2.0
mols2grid>=1.0.0

# Deep Learning (optional)
tensorflow>=2.8.0
torch>=1.12.0
torch-geometric>=2.1.0

# Jupyter and Development
jupyter>=1.0.0
ipywidgets>=7.6.0
tqdm>=4.64.0

# Utilities
joblib>=1.1.0
pickle5>=0.0.11

# Web Interface (optional)
streamlit>=1.25.0

# Report Generation
reportlab>=3.6.0
Pillow>=9.0.0